compile properly with all warnings enabled (some size_t and ssize_t used, pragmas added for global constructors/destructors)

+++ packages/hal/arch/synthetic/osx/devices/debug/include/DebugImplementation.h
@@ -29,7 +29,7 @@ namespace os
           static void
           putNewLine(void);
 
-          static int
+          static ssize_t
           putBytes(const void* cpBuf, size_t n);
 
         };
